블랙홀을 장애물화, 블랙홀 페이즈 추가 대폭발 페이즈 추가 거대 유성 출현 페이즈를 무조건 거대유성이 되게 함 시간에 따라 운석 갯수 추가하지 않게 함 스코어 업데이트때 큐브콜렉터 보너스 정상 적용

+++ assets/01_Scripts/20_InGame/Others/DangerousEMPField.cs
@@ -4,7 +4,7 @@ using System.Collections;
 public class DangerousEMPField : MonoBehaviour {
   PlayerMover player;
   DangerousEMPManager dem;
-  int maxScale;
+  float maxScale;
   int rotatingSpeed;
   float enlargeDuration;
   float stayDuration;
